Migrate to TypeScript, ship ESM-only, modernise the build
- Source: src/*.js → src/*.ts with full type annotations. Generic Data
  parameter on Route, Router, RouteDefinition so consumers get typed
  metadata. Internal imports use .ts extensions and tsc rewrites them
  to .js on emit (allowImportingTsExtensions + rewriteRelativeImport-
  Extensions).
- Output: ESM only. Drop CJS dual-publish. dist/ is flat (no /cjs,
  /esm subdirs); .js and .d.ts sit side by side. Modern exports map,
  package "type": "module", "sideEffects": false, files whitelist.
- Build: drop swc and the custom build.js. tsc alone handles transpile
  + declarations. Removes @swc/cli, @swc/core, @swc-node/register,
  execa, tasks/.
- Tests: ava with Node's native TypeScript stripping — no tsx, no
  loaders, no plugins. Just extensions: { ts: 'module' }.
- Drop the index.d.ts shim at the root (now generated to dist/).
- Drop the minimatch override (no longer needed; ava and c8 both
  pull minimatch@10 directly).
- Source cleanups while here:
  - history: queueMicrotask in place of the (globalThis as ...).set-
    Immediate cast for SSR; drop the IE9-era \!history.pushState hash
    fallback; getHash uses location.hash.slice(1).
  - match: matchOne returns MatchedRoute | undefined (was | false);
    drop the ret sentinel for early returns; modernise the flag
    extraction to ?.[0] ?? ''.

Engines bumped to >=18 for ESM. Published tarball: 51 files / 342 kB
→ 18 files / 28 kB.
