--- .npmignore
@@ -1,5 +1,0 @@
-.DS_Store
-.nyc_output
-coverage
-docs/public
-docs/resources/_gen
