--- .swc-cjs
@@ -1,11 +1,0 @@
-{
-  "module": {
-    "type": "commonjs"
-  },
-  "env": {
-    "targets": [
-      "last 2 versions",
-      "IE >= 9"
-    ]
-  }
-}
