--- .swc-esm
@@ -1,1 +1,0 @@
-{}
