+++ src/history.ts
@@ -1,29 +1,40 @@
-export function createHistory(options = {}) {
-  const sync = options.sync
-  let mode = options.mode || 'history'
-  let raf
-  let listener
+export type Mode = 'history' | 'hash' | 'memory'
 
-  const memory = []
+export interface History {
+  listen(onChange: (url: string) => void): () => void
+  getUrl(): string
+  push(url: string): void
+  replace(url: string): void
+}
+
+export interface CreateHistoryOptions {
+  mode?: Mode
+  sync?: boolean
+}
+
+export function createHistory(options: CreateHistoryOptions = {}): History {
+  const sync = options.sync
+  let mode: Mode = options.mode || 'history'
+  let raf: (cb: () => void) => void
+  let listener: ((url: string) => void) | null = null
+
+  const memory: string[] = []
 
   if (typeof window === 'undefined') {
     mode = 'memory'
-    raf = sync ? (cb) => cb() : global.setImmediate
+    raf = sync ? (cb) => cb() : queueMicrotask
   } else {
     raf = sync ? (cb) => cb() : requestAnimationFrame
-    if (mode === 'history' && !history.pushState) {
-      mode = 'hash'
-    }
   }
 
   function emit() {
     if (listener) listener(getUrl())
   }
 
-  function listen(onChange) {
+  function listen(onChange: (url: string) => void) {
     if (listener) throw new Error('Already listening')
     listener = onChange
-    let off
+    let off: (() => void) | undefined
     if (mode !== 'memory') {
       off = on(window, mode === 'history' ? 'popstate' : 'hashchange', emit)
       raf(emit)
@@ -34,7 +45,7 @@ export function createHistory(options = {}) {
     }
   }
 
-  function go(url, replace) {
+  function go(url: string, replace?: boolean) {
     url = url.replace(/^\/?#?\/?/, '/').replace(/\/$/, '') || '/'
     if (mode === 'history') {
       history[replace ? 'replaceState' : 'pushState']({}, '', url)
@@ -51,7 +62,7 @@ export function createHistory(options = {}) {
     }
   }
 
-  function getUrl() {
+  function getUrl(): string {
     if (mode === 'memory') {
       return memory[memory.length - 1]
     }
@@ -61,21 +72,16 @@ export function createHistory(options = {}) {
       return hash === '' ? '/' : hash
     }
 
-    if (mode === 'history') {
-      let url = location.pathname + location.search
-      if (hash !== '') {
-        url += '#' + hash
-      }
-
-      return url
+    let url = location.pathname + location.search
+    if (hash !== '') {
+      url += '#' + hash
     }
+
+    return url
   }
 
-  // Gets the true hash value. Cannot use location.hash directly due to bug
-  // in Firefox where location.hash will always be decoded.
   function getHash() {
-    const match = location.href.match(/#(.*)$/)
-    return match ? match[1] : ''
+    return location.hash.slice(1)
   }
 
   return {
@@ -90,7 +96,7 @@ export function createHistory(options = {}) {
   }
 }
 
-function on(el, type, fn) {
+function on(el: Window, type: string, fn: () => void) {
   el.addEventListener(type, fn, false)
   return function off() {
     el.removeEventListener(type, fn, false)
