--- src/index.js
@@ -1,3 +1,0 @@
-export { createRouter, merge } from './router'
-export { createHistory } from './history'
-export { qs } from './qs'
