+++ src/index.ts
@@ -1,0 +1,7 @@
+export { qs } from './qs.ts'
+export { createHistory } from './history.ts'
+export { createRouter, merge } from './router.ts'
+
+export type { Qs } from './qs.ts'
+export type { Mode, History, CreateHistoryOptions } from './history.ts'
+export type { RouterOptions, Route, NavigateTarget, To, Redirect, RouteDefinition, Router } from './router.ts'
